refactor(server): move ErrorInfoList job to dedicated files
Refactored ErrorInfoList job by renaming and moving its code to errorinfolistjob.{h,cpp}. Updated all includes and CMakeLists.txt references. Separated its tests from testguicommchannel.cpp into a new testerrorjobs.cpp file for better modularity. No functional changes; improves code organization and maintainability.
